# api/swagger.yml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Kurirmoo
schemes: [http]
consumes: [application/json, multipart/form-data]
produces: [application/json]
securityDefinitions:
  key:
    type: apiKey
    in: header
    name: Authorization
  hasRoleSuperAdmin:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  /api/v1/trucks:
    post:
      tags: [trucks]
      summary: "add truck"
      description: "Add new truck"
      operationId: addTruck
      security: []
      parameters:
        - name: data
          in: body
          required: true
          schema:
            type: object
            required:
              - truck_kind
              - brand
              - truck_type
              - length
              - width
              - height
              - capacity
            properties:
              truck_kind:
                type: string
              brand:
                type: string
              truck_type:
                type: string
              length:
                type: integer
              width:
                type: integer
              height:
                type: integer
              capacity:
                type: integer
      responses:
        201:
          description: "Success added"
          schema:
            $ref: "#/definitions/Success"
        400:
          $ref: "#/responses/BadRequest"
        default:
          description: error
          schema:
            $ref: "#/definitions/Error"

  ## data multipliers (data pengali)
...
  /api/v1/routes/{id}:
    get:
      tags: [getRouteAndCityPasseds]
      summary: get route and city passeds
      description: Return route and city passeds
      operationId: getRouteAndCityPasseds
      security: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              origin_city:
                type: string
              destination_city:
                type: string
              city_passeds:
                type: array
                items:
                  type: string
        400:
          $ref: "#/responses/BadRequest"
        404:
          $ref: "#/responses/NotFound"
        500:
          $ref: "#/responses/InternalServerError"
        default:
          description: error
          schema:
            $ref: "#/definitions/Error"
      produces:
        - application/json
definitions:
...
  Error:
    type: object
    properties:
      code:
        type: integer
      message:
        type: string
        example: "error"
  Success:
    type: object
    properties:
      message:
        type: string
...
responses:
  NotFound:
    description: The specified resource was not found
    schema:
      $ref: "#/definitions/Error"
...
  BadRequest:
    description: Bad Request
    schema:
      $ref: "#/definitions/Error"
...
  InternalServerError:
    description: Internal server error
    schema:
      $ref: "#/definitions/Error"
...